backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
openai==1.12.0
pandas==2.2.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.1
